requirements.txt
pandas
textblob
psycopg2-binary
streamlit
plotly
schedule
praw
wordcloud
flask